+++ src/TrelloDotNet/Model/EmojiResponse.cs
@@ -1,0 +1,12 @@
+using System.Collections.Generic;
+using System.Text.Json.Serialization;
+
+namespace TrelloDotNet.Model
+{
+    internal class EmojiResponse
+    {
+        [JsonPropertyName("trello")]
+        [JsonInclude]
+        public List<Emoji> Trello { get; set; }
+    }
+}